+++ src/directives/index.js
@@ -3,4 +3,5 @@ module.exports = ngModule => {
   require('./formly-dynamic-name')(ngModule);
   require('./formly-field')(ngModule);
   require('./formly-form')(ngModule);
+  require('./formly-focus')(ngModule);
 };